Worked on the camera scripting and scripting namespaces.
I also watched the first Narnia film while doing it.
I didn't like those films as a child, and I was actually very surprised as to how much more I disliked them as an adult.
It all just feels wrong, like an off brand Lord of the Rings.
The setting is un-interesting, the stakes don't seem that high, there are blatant oversights in the writing.
Like who thought giving a child a magic bottle of magic heal-all plot cheat juice would be a good idea.
None of the non-evil characters die. Where are the stakes!
Death is an important part of story telling, and Aslan just comes back to life like a top geezer but why does that have to happen?
It does nothing but destroy the emotion built up from his death scene, which is otherwise pretty good.
Narnia doesn't feel like a wide, expansive universe, it just feels like some scatty field with a lion.
There's just nothing going on that I care to care about.

We also watched the 2016 jungle book, which I thought was actually alright.